+++ R/RcppExports.R
@@ -28,14 +28,9 @@ cpp_rcbbinom <- function(n, size, alpha, beta, p_tol, p_max_iter, p_prec, root_t
 #' \code{boost::math::tools::promote_args} (e.g. for \code{double} input,
 #' usually use the epsilon for \code{long double}).
 #' @param max_iter Integer (1L) as iteration limit.
-#' @param prec Numeric (1L) as precision level during computation,
-#' a.k.a the number of precise digits defined in \code{boost::multiprecision::cpp_bin_float}.
-#' For level 0, \code{double} precision is used. For level 20, the precision is slightly
-#' higher than \code{long double}
-#' (\code{.Machine[["longdouble.eps"]] = `r format(.Machine[["longdouble.eps"]], scientific = TRUE)`}).
-#' Higher levels include 25/30/35/.../50. The higher the accuracy,
-#' the slower the computation. For an input other than these values, the lowest level
-#' greater than the input is used, up to 50.
+#' @param prec \code{NULL} or (unsigned) integer (1L) as precision level during computation,
+#' a.k.a the number of precise digits defined in \code{boost::multiprecision::mpfr_float}.
+#' For \code{NULL}, double precision (default) is used.
 #' @param check_mode Logical (1L) indicating whether the mode of \code{x}
 #' should be checked for obvious convergence failures.
 #' @param log Logical (1L) indicating whether result is given as log(result).
@@ -54,7 +49,7 @@ cpp_rcbbinom <- function(n, size, alpha, beta, p_tol, p_max_iter, p_prec, root_t
 #'
 #' @examples
 #' gen_hypergeo(U = c(1.1, 0.2, 0.3), L = c(10.1, 4 * pi), x = 1,
-#'              tol = NULL, max_iter = 10000L, prec = 20, check_mode = TRUE, log = FALSE)
+#'              tol = NULL, max_iter = 10000L, prec = NULL, check_mode = TRUE, log = FALSE)
 gen_hypergeo <- function(U, L, x, tol, max_iter, prec, check_mode, log) {
     .Call(`_cbbinom_gen_hypergeo`, U, L, x, tol, max_iter, prec, check_mode, log)
 }