+++ src/RcppExports.cpp
@@ -14,7 +14,7 @@ Rcpp::Rostream<false>& Rcpp::Rcerr = Rcpp::Rcpp_cerr_get();
 #endif
 
 // cpp_pcbbinom
-NumericVector cpp_pcbbinom(const NumericVector& q, const NumericVector& size, const NumericVector& alpha, const NumericVector& beta, const bool& lower_tail, const bool& log_p, const Nullable<NumericVector>& tol, const IntegerVector& max_iter, const NumericVector& prec);
+NumericVector cpp_pcbbinom(const NumericVector& q, const NumericVector& size, const NumericVector& alpha, const NumericVector& beta, const bool& lower_tail, const bool& log_p, const Nullable<NumericVector>& tol, const IntegerVector& max_iter, const Nullable<IntegerVector>& prec);
 static SEXP _cbbinom_cpp_pcbbinom_try(SEXP qSEXP, SEXP sizeSEXP, SEXP alphaSEXP, SEXP betaSEXP, SEXP lower_tailSEXP, SEXP log_pSEXP, SEXP tolSEXP, SEXP max_iterSEXP, SEXP precSEXP) {
 BEGIN_RCPP
     Rcpp::RObject rcpp_result_gen;
@@ -26,7 +26,7 @@ BEGIN_RCPP
     Rcpp::traits::input_parameter< const bool& >::type log_p(log_pSEXP);
     Rcpp::traits::input_parameter< const Nullable<NumericVector>& >::type tol(tolSEXP);
     Rcpp::traits::input_parameter< const IntegerVector& >::type max_iter(max_iterSEXP);
-    Rcpp::traits::input_parameter< const NumericVector& >::type prec(precSEXP);
+    Rcpp::traits::input_parameter< const Nullable<IntegerVector>& >::type prec(precSEXP);
     rcpp_result_gen = Rcpp::wrap(cpp_pcbbinom(q, size, alpha, beta, lower_tail, log_p, tol, max_iter, prec));
     return rcpp_result_gen;
 END_RCPP_RETURN_ERROR
@@ -56,7 +56,7 @@ RcppExport SEXP _cbbinom_cpp_pcbbinom(SEXP qSEXP, SEXP sizeSEXP, SEXP alphaSEXP,
     return rcpp_result_gen;
 }
 // cpp_qcbbinom
-NumericVector cpp_qcbbinom(const NumericVector& p, const NumericVector& size, const NumericVector& alpha, const NumericVector& beta, const bool& lower_tail, const bool& log_p, const Nullable<NumericVector>& p_tol, const IntegerVector& p_max_iter, const NumericVector& p_prec, const NumericVector& root_tol, const IntegerVector& root_max_iter);
+NumericVector cpp_qcbbinom(const NumericVector& p, const NumericVector& size, const NumericVector& alpha, const NumericVector& beta, const bool& lower_tail, const bool& log_p, const Nullable<NumericVector>& p_tol, const IntegerVector& p_max_iter, const Nullable<IntegerVector>& p_prec, const NumericVector& root_tol, const IntegerVector& root_max_iter);
 static SEXP _cbbinom_cpp_qcbbinom_try(SEXP pSEXP, SEXP sizeSEXP, SEXP alphaSEXP, SEXP betaSEXP, SEXP lower_tailSEXP, SEXP log_pSEXP, SEXP p_tolSEXP, SEXP p_max_iterSEXP, SEXP p_precSEXP, SEXP root_tolSEXP, SEXP root_max_iterSEXP) {
 BEGIN_RCPP
     Rcpp::RObject rcpp_result_gen;
@@ -68,7 +68,7 @@ BEGIN_RCPP
     Rcpp::traits::input_parameter< const bool& >::type log_p(log_pSEXP);
     Rcpp::traits::input_parameter< const Nullable<NumericVector>& >::type p_tol(p_tolSEXP);
     Rcpp::traits::input_parameter< const IntegerVector& >::type p_max_iter(p_max_iterSEXP);
-    Rcpp::traits::input_parameter< const NumericVector& >::type p_prec(p_precSEXP);
+    Rcpp::traits::input_parameter< const Nullable<IntegerVector>& >::type p_prec(p_precSEXP);
     Rcpp::traits::input_parameter< const NumericVector& >::type root_tol(root_tolSEXP);
     Rcpp::traits::input_parameter< const IntegerVector& >::type root_max_iter(root_max_iterSEXP);
     rcpp_result_gen = Rcpp::wrap(cpp_qcbbinom(p, size, alpha, beta, lower_tail, log_p, p_tol, p_max_iter, p_prec, root_tol, root_max_iter));
@@ -100,7 +100,7 @@ RcppExport SEXP _cbbinom_cpp_qcbbinom(SEXP pSEXP, SEXP sizeSEXP, SEXP alphaSEXP,
     return rcpp_result_gen;
 }
 // cpp_dcbbinom
-NumericVector cpp_dcbbinom(const NumericVector& x, const NumericVector& size, const NumericVector& alpha, const NumericVector& beta, const bool& log, const Nullable<NumericVector>& tol, const IntegerVector& max_iter, const NumericVector& prec);
+NumericVector cpp_dcbbinom(const NumericVector& x, const NumericVector& size, const NumericVector& alpha, const NumericVector& beta, const bool& log, const Nullable<NumericVector>& tol, const IntegerVector& max_iter, const Nullable<IntegerVector>& prec);
 static SEXP _cbbinom_cpp_dcbbinom_try(SEXP xSEXP, SEXP sizeSEXP, SEXP alphaSEXP, SEXP betaSEXP, SEXP logSEXP, SEXP tolSEXP, SEXP max_iterSEXP, SEXP precSEXP) {
 BEGIN_RCPP
     Rcpp::RObject rcpp_result_gen;
@@ -111,7 +111,7 @@ BEGIN_RCPP
     Rcpp::traits::input_parameter< const bool& >::type log(logSEXP);
     Rcpp::traits::input_parameter< const Nullable<NumericVector>& >::type tol(tolSEXP);
     Rcpp::traits::input_parameter< const IntegerVector& >::type max_iter(max_iterSEXP);
-    Rcpp::traits::input_parameter< const NumericVector& >::type prec(precSEXP);
+    Rcpp::traits::input_parameter< const Nullable<IntegerVector>& >::type prec(precSEXP);
     rcpp_result_gen = Rcpp::wrap(cpp_dcbbinom(x, size, alpha, beta, log, tol, max_iter, prec));
     return rcpp_result_gen;
 END_RCPP_RETURN_ERROR
@@ -141,7 +141,7 @@ RcppExport SEXP _cbbinom_cpp_dcbbinom(SEXP xSEXP, SEXP sizeSEXP, SEXP alphaSEXP,
     return rcpp_result_gen;
 }
 // cpp_rcbbinom
-NumericVector cpp_rcbbinom(const int& n, const NumericVector& size, const NumericVector& alpha, const NumericVector& beta, const Nullable<NumericVector>& p_tol, const IntegerVector& p_max_iter, const NumericVector& p_prec, const NumericVector& root_tol, const IntegerVector& root_max_iter);
+NumericVector cpp_rcbbinom(const int& n, const NumericVector& size, const NumericVector& alpha, const NumericVector& beta, const Nullable<NumericVector>& p_tol, const IntegerVector& p_max_iter, const Nullable<IntegerVector>& p_prec, const NumericVector& root_tol, const IntegerVector& root_max_iter);
 static SEXP _cbbinom_cpp_rcbbinom_try(SEXP nSEXP, SEXP sizeSEXP, SEXP alphaSEXP, SEXP betaSEXP, SEXP p_tolSEXP, SEXP p_max_iterSEXP, SEXP p_precSEXP, SEXP root_tolSEXP, SEXP root_max_iterSEXP) {
 BEGIN_RCPP
     Rcpp::RObject rcpp_result_gen;
@@ -151,7 +151,7 @@ BEGIN_RCPP
     Rcpp::traits::input_parameter< const NumericVector& >::type beta(betaSEXP);
     Rcpp::traits::input_parameter< const Nullable<NumericVector>& >::type p_tol(p_tolSEXP);
     Rcpp::traits::input_parameter< const IntegerVector& >::type p_max_iter(p_max_iterSEXP);
-    Rcpp::traits::input_parameter< const NumericVector& >::type p_prec(p_precSEXP);
+    Rcpp::traits::input_parameter< const Nullable<IntegerVector>& >::type p_prec(p_precSEXP);
     Rcpp::traits::input_parameter< const NumericVector& >::type root_tol(root_tolSEXP);
     Rcpp::traits::input_parameter< const IntegerVector& >::type root_max_iter(root_max_iterSEXP);
     rcpp_result_gen = Rcpp::wrap(cpp_rcbbinom(n, size, alpha, beta, p_tol, p_max_iter, p_prec, root_tol, root_max_iter));
@@ -183,7 +183,7 @@ RcppExport SEXP _cbbinom_cpp_rcbbinom(SEXP nSEXP, SEXP sizeSEXP, SEXP alphaSEXP,
     return rcpp_result_gen;
 }
 // gen_hypergeo
-double gen_hypergeo(const NumericVector& U, const NumericVector& L, const double& x, const Nullable<NumericVector>& tol, const R_xlen_t& max_iter, const double& prec, const bool& check_mode, const bool& log);
+double gen_hypergeo(const NumericVector& U, const NumericVector& L, const double& x, const Nullable<NumericVector>& tol, const R_xlen_t& max_iter, const Nullable<IntegerVector>& prec, const bool& check_mode, const bool& log);
 static SEXP _cbbinom_gen_hypergeo_try(SEXP USEXP, SEXP LSEXP, SEXP xSEXP, SEXP tolSEXP, SEXP max_iterSEXP, SEXP precSEXP, SEXP check_modeSEXP, SEXP logSEXP) {
 BEGIN_RCPP
     Rcpp::RObject rcpp_result_gen;
@@ -192,7 +192,7 @@ BEGIN_RCPP
     Rcpp::traits::input_parameter< const double& >::type x(xSEXP);
     Rcpp::traits::input_parameter< const Nullable<NumericVector>& >::type tol(tolSEXP);
     Rcpp::traits::input_parameter< const R_xlen_t& >::type max_iter(max_iterSEXP);
-    Rcpp::traits::input_parameter< const double& >::type prec(precSEXP);
+    Rcpp::traits::input_parameter< const Nullable<IntegerVector>& >::type prec(precSEXP);
     Rcpp::traits::input_parameter< const bool& >::type check_mode(check_modeSEXP);
     Rcpp::traits::input_parameter< const bool& >::type log(logSEXP);
     rcpp_result_gen = Rcpp::wrap(gen_hypergeo(U, L, x, tol, max_iter, prec, check_mode, log));
@@ -228,11 +228,11 @@ RcppExport SEXP _cbbinom_gen_hypergeo(SEXP USEXP, SEXP LSEXP, SEXP xSEXP, SEXP t
 static int _cbbinom_RcppExport_validate(const char* sig) { 
     static std::set<std::string> signatures;
     if (signatures.empty()) {
-        signatures.insert("NumericVector(*cpp_pcbbinom)(const NumericVector&,const NumericVector&,const NumericVector&,const NumericVector&,const bool&,const bool&,const Nullable<NumericVector>&,const IntegerVector&,const NumericVector&)");
-        signatures.insert("NumericVector(*cpp_qcbbinom)(const NumericVector&,const NumericVector&,const NumericVector&,const NumericVector&,const bool&,const bool&,const Nullable<NumericVector>&,const IntegerVector&,const NumericVector&,const NumericVector&,const IntegerVector&)");
-        signatures.insert("NumericVector(*cpp_dcbbinom)(const NumericVector&,const NumericVector&,const NumericVector&,const NumericVector&,const bool&,const Nullable<NumericVector>&,const IntegerVector&,const NumericVector&)");
-        signatures.insert("NumericVector(*cpp_rcbbinom)(const int&,const NumericVector&,const NumericVector&,const NumericVector&,const Nullable<NumericVector>&,const IntegerVector&,const NumericVector&,const NumericVector&,const IntegerVector&)");
-        signatures.insert("double(*gen_hypergeo)(const NumericVector&,const NumericVector&,const double&,const Nullable<NumericVector>&,const R_xlen_t&,const double&,const bool&,const bool&)");
+        signatures.insert("NumericVector(*cpp_pcbbinom)(const NumericVector&,const NumericVector&,const NumericVector&,const NumericVector&,const bool&,const bool&,const Nullable<NumericVector>&,const IntegerVector&,const Nullable<IntegerVector>&)");
+        signatures.insert("NumericVector(*cpp_qcbbinom)(const NumericVector&,const NumericVector&,const NumericVector&,const NumericVector&,const bool&,const bool&,const Nullable<NumericVector>&,const IntegerVector&,const Nullable<IntegerVector>&,const NumericVector&,const IntegerVector&)");
+        signatures.insert("NumericVector(*cpp_dcbbinom)(const NumericVector&,const NumericVector&,const NumericVector&,const NumericVector&,const bool&,const Nullable<NumericVector>&,const IntegerVector&,const Nullable<IntegerVector>&)");
+        signatures.insert("NumericVector(*cpp_rcbbinom)(const int&,const NumericVector&,const NumericVector&,const NumericVector&,const Nullable<NumericVector>&,const IntegerVector&,const Nullable<IntegerVector>&,const NumericVector&,const IntegerVector&)");
+        signatures.insert("double(*gen_hypergeo)(const NumericVector&,const NumericVector&,const double&,const Nullable<NumericVector>&,const R_xlen_t&,const Nullable<IntegerVector>&,const bool&,const bool&)");
     }
     return signatures.find(sig) != signatures.end();
 }